summarize_results.sh: Check if best.txt exists

+++ eval_scripts/summarize_results.sh
@@ -37,7 +37,9 @@ function create_graph
 	echo }
 }
 
-rm tmp/best.txt
+if [ -f tmp/best.txt ]; then
+	rm tmp/best.txt
+fi
 grep -v '#' ../src/models/test_models.txt|tr -d '!' >tmp/models.tmp
 for m in $(cut -f 1 -d ' ' tmp/models.tmp)
 do
